+++ examples/cardano-sl.nix
@@ -1,3 +1,3 @@
 let inherit (import ../. {}) sources nixpkgsArgs;
-    pkgs = import sources.nixpkgs-default nixpkgsArgs;
+    pkgs = import sources.nixpkgs nixpkgsArgs;
 in (pkgs.haskell-nix.stackProject (import ./cardano-sl-args.nix)).cardano-sl.components.all