+++ scripts/update-hackage.nix
@@ -1,7 +1,6 @@
 { stdenv, writeScript, coreutils, glibc, git, openssh
-, nix-tools, cabal-install, nix-prefetch-git
+, internal-nix-tools, internal-cabal-install, nix-prefetch-git
 , gawk, bash, curl, findutils
-, compiler-nix-name
 , update-index-state-hashes }@args:
 
 import ./update-external.nix
@@ -30,6 +29,6 @@ import ./update-external.nix
 
     cd hackage.nix
 
-    ${update-index-state-hashes compiler-nix-name}/bin/update-index-state-hashes > index-state-hashes.nix
+    ${update-index-state-hashes}/bin/update-index-state-hashes > index-state-hashes.nix
   '';
 }