+++ test/index-state/default.nix
@@ -6,8 +6,7 @@ with stdenv.lib;
 let
   # The hackage-security 0.6.0.1 was uploaded at 2020-04-06T20:54:35Z
   # See https://hackage.haskell.org/package/hackage-security-0.6.0.1
-  version-used-at = index-state: (tool "cabal" {
-      inherit compiler-nix-name;
+  version-used-at = index-state: (tool compiler-nix-name "cabal" {
       version = "3.2.0.0";
       inherit index-state;
       cabalProject = ''